requirements.txt
accelerate==1.10.1
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
psutil==7.1.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.15.0
urllib3==2.5.0
